refactor: move TargetButton component and update imports in skill maker files
