+++ app/src/main/java/io/github/fate_grand_automata/ui/skill_maker/special/SkillMakerChangeNpType2.kt
@@ -8,9 +8,7 @@ import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.fillMaxHeight
 import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.layout.width
 import androidx.compose.material3.Button
 import androidx.compose.material3.ButtonDefaults
 import androidx.compose.material3.MaterialTheme
@@ -22,20 +20,18 @@ import androidx.compose.runtime.remember
 import androidx.compose.runtime.setValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.res.colorResource
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.text.style.TextDecoration
-import androidx.compose.ui.text.style.TextOverflow
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.Dp
 import androidx.compose.ui.unit.dp
-import androidx.compose.ui.unit.sp
 import io.github.fate_grand_automata.R
 import io.github.fate_grand_automata.ui.FGATheme
 import io.github.fate_grand_automata.ui.FGATitle
 import io.github.fate_grand_automata.ui.skill_maker.utils.ChangeNp2Type
+import io.github.fate_grand_automata.ui.skill_maker.utils.TargetButton
 import io.github.fate_grand_automata.ui.skill_maker.utils.stringRes
 import io.github.fate_grand_automata.ui.skill_maker.utils.targetAStringRes
 import io.github.fate_grand_automata.ui.skill_maker.utils.targetBStringRes
@@ -139,30 +135,6 @@ fun SkillMakerChangeNpType2(
     }
 }
 
-@Composable
-fun TargetButton(
-    onClick: () -> Unit,
-    color: Color,
-    enabled: Boolean = true,
-    text: String
-) {
-    Button(
-        onClick = onClick,
-        colors = ButtonDefaults.buttonColors(containerColor = color),
-        modifier = Modifier
-            .height(75.dp)
-            .width(120.dp),
-        enabled = enabled
-    ) {
-        Text(
-            text,
-            color = Color.White,
-            textAlign = TextAlign.Center,
-            overflow = TextOverflow.Clip,
-            fontSize = 17.sp
-        )
-    }
-}
 
 @Preview(name = "Light Mode", widthDp = 600, heightDp = 300)
 @Preview(name = "Dark Mode", widthDp = 600, heightDp = 300, uiMode = Configuration.UI_MODE_NIGHT_YES)
