+++ app/src/main/java/io/github/fate_grand_automata/ui/skill_maker/special/SkillMakerChangeNpType3.kt
@@ -32,6 +32,7 @@ import io.github.fate_grand_automata.scripts.models.ServantTarget
 import io.github.fate_grand_automata.ui.FGATheme
 import io.github.fate_grand_automata.ui.FGATitle
 import io.github.fate_grand_automata.ui.skill_maker.utils.ChangeNp3Type
+import io.github.fate_grand_automata.ui.skill_maker.utils.TargetButton
 import io.github.fate_grand_automata.ui.skill_maker.utils.stringRes
 import io.github.fate_grand_automata.ui.skill_maker.utils.targetAStringRes
 import io.github.fate_grand_automata.ui.skill_maker.utils.targetBStringRes
